Use previously selected notebook for a new note
If the user has filtered the sidebar by a notebook before they click the
new note button, the notebook is selected by default for the note.

The filter and query are included in the URI for the `notes/add` route.
The note form will use these two values to figure out which notebook to
pre-select in case the currently edited note doesn't have a notebook
assigned.

Implements #271