Fix a translation bug with "You have unsaved changes."
The "You have unsaved changes." sentence in the cancel dialog was never translated. This was due to the dot in the sentence in the translation files. In JSON files, this dot is a call for a subclass, not the end of a sentence. Because of that, the translation was never found. I deleted the dot in the files where a new cancel dialog is created and in all translation files.