Fix the variables in the translation files
Fixes a problem with the use of variables in the translation json files.
i18n uses {{variable}} to access variables, but in the files all
variables were marked with __variable__ . Because of this, no variables
were loaded and just the string between within the underscores was
displayed, not the variable itself. This has been fixed.