Translate profile name in Settings-Profile
Now the "profile name" string will be translated. Although the i18n mark
was set in the profiles.html, the "Profile name" was missing in the
translation files. So I changed the i18n call to 'profile.profile name',
updated the translation in the Russian and Turkish translation and added
the translation to the English and German file.