Show notification for wrong password
Fixes #520

As written in the issue there is a trigger when the password is wrong.
But it wasn't implemented in the view and the call had to be made by
self, not this. The method that is triggered now shakes the
button and clears the password form. The shake function was implemented
in this function because I didn't want to use the whole jQuery-UI
package just for this.