Don't save "wrong" tags anymore
Fixes #493

"Wrong" means saved tags due to the auto save feature which saved the tags,
even if you were not finished writing them or deleted them afterwards. I fixed this by
creating the flag "saveTags", which is false when the note is auto saved
and true when it's really saved. So now tags are only saved when you
save the note on your own.

Thanks to @wwebfor for giving me a new way of seeing this problem and
thus making it much easier to solve for me.